[Move] sideBar, sideView and editor into `parts` dir
