+++ src/code/browser/renderer.ts
@@ -1,5 +1,5 @@
 import "src/styles/index.scss";
-import "src/code/browser/workbench/workspace/editor/editor";
+import "src/code/browser/workbench/parts/workspace/editor/editor";
 import { Workbench } from "src/code/browser/workbench/workbench";
 import { registerBrowserDefaultConfiguration } from "src/code/platform/configuration/browser/configuration.register";
 import { rendererServiceRegistrations } from "src/code/browser/service.register";
