+++ src/code/browser/service.register.ts
@@ -6,12 +6,12 @@ import { IKeyboardService, keyboardService } from "src/code/browser/service/keyb
 import { BrowserDialogService, IDialogService } from 'src/code/platform/dialog/browser/browserDialogService';
 import { IShortcutService, ShortcutService } from 'src/code/browser/service/shortcut/shortcutService';
 import { IThemeService, ThemeService } from 'src/code/browser/service/theme/themeService';
-import { SideBar, ISideBarService } from 'src/code/browser/workbench/sideBar/sideBar';
-import { IWorkspaceService, WorkspaceComponent } from 'src/code/browser/workbench/workspace/workspace';
-import { SideViewService, ISideViewService } from 'src/code/browser/workbench/sideView/sideView';
+import { SideBar, ISideBarService } from 'src/code/browser/workbench/parts/sideBar/sideBar';
+import { IWorkspaceService, WorkspaceComponent } from 'src/code/browser/workbench/parts/workspace/workspace';
+import { SideViewService, ISideViewService } from 'src/code/browser/workbench/parts/sideView/sideView';
 import { ExplorerTreeService, IExplorerTreeService } from 'src/code/browser/service/explorerTree/explorerTreeService';
 import { ContextService, IContextService } from 'src/code/platform/context/common/contextService';
-import { ContextMenuService, IContextMenuService } from "src/code/browser/service/contextMenu/contextMenuService";
+// import { ContextMenuService, IContextMenuService } from "src/code/browser/service/contextMenu/contextMenuService";
 
 /*******************************************************************************
  * Registraion for desktop browser-side non-important microservices.
@@ -34,7 +34,7 @@ export function rendererServiceRegistrations(): void {
     registerSingleton(IKeyboardScreenCastService, new ServiceDescriptor(KeyboardScreenCastService));
     registerSingleton(IThemeService, new ServiceDescriptor(ThemeService)); // TODO: themeService
     registerSingleton(IExplorerTreeService, new ServiceDescriptor(ExplorerTreeService));
-    registerSingleton(IContextMenuService, new ServiceDescriptor(ContextMenuService));
+    // registerSingleton(IContextMenuService, new ServiceDescriptor(ContextMenuService));
 
     // utilities && tools
     registerSingleton(IContextService, new ServiceDescriptor(ContextService));
