+++ src/code/browser/workbench/contrib/explorer/explorerService.ts
@@ -1,6 +1,6 @@
 import { Register } from "src/base/common/event";
 import { URI } from "src/base/common/file/uri";
-import { ISideView } from "src/code/browser/workbench/sideView/sideView";
+import { ISideView } from "src/code/browser/workbench/parts/sideView/sideView";
 
 export const ExplorerViewID = 'explorer-view';
