+++ src/code/browser/workbench/parts/sideBar/media/sideBar.scss
