+++ src/code/browser/workbench/parts/sideBar/sideBarButton.ts
