+++ src/code/browser/workbench/parts/sideView/media/sideView.scss
