+++ src/code/browser/workbench/parts/sideView/sideViewTitle.ts
