+++ src/code/browser/workbench/parts/workspace/editor/media/editor.scss
