+++ src/code/browser/workbench/parts/workspace/media/workspace.scss
