+++ src/code/browser/workbench/parts/workspace/titleBar/media/titleBar.scss
