+++ src/code/browser/workbench/parts/workspace/titleBar/media/windowBar.scss
