+++ src/code/browser/workbench/parts/workspace/titleBar/windowButton.ts
