+++ src/code/browser/workbench/workbench.ts
@@ -1,6 +1,5 @@
 import 'src/code/browser/workbench/media/workbench.scss';
 import { IInstantiationService } from "src/code/platform/instantiation/common/instantiation";
-import { ServiceDescriptor } from "src/code/platform/instantiation/common/descriptor";
 import { IComponentService } from "src/code/browser/service/component/componentService";
 import { WorkbenchLayout } from "src/code/browser/workbench/layout";
 import { IWorkbenchService } from "src/code/browser/service/workbench/workbenchService";
@@ -8,9 +7,9 @@ import { IKeyboardScreenCastService } from "src/code/browser/service/keyboard/ke
 import { IConfigService } from "src/code/platform/configuration/common/abstractConfigService";
 import { BuiltInConfigScope } from "src/code/platform/configuration/common/configRegistrant";
 import { IThemeService } from "src/code/browser/service/theme/themeService";
-import { ISideBarService } from "src/code/browser/workbench/sideBar/sideBar";
-import { ISideViewService } from "src/code/browser/workbench/sideView/sideView";
-import { IWorkspaceService } from "src/code/browser/workbench/workspace/workspace";
+import { ISideBarService } from "src/code/browser/workbench/parts/sideBar/sideBar";
+import { ISideViewService } from "src/code/browser/workbench/parts/sideView/sideView";
+import { IWorkspaceService } from "src/code/browser/workbench/parts/workspace/workspace";
 import { Disposable } from 'src/base/common/dispose';
 import { IContextService } from 'src/code/platform/context/common/contextService';
 import { IContextKey } from 'src/code/platform/context/common/contextKey';
